adding new trips works

+++ src/main/java/com/zak/cruise/dto/CruiseDTO.java
@@ -6,11 +6,15 @@ import com.zak.cruise.entity.Ship;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
+import org.hibernate.validator.constraints.Length;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.stereotype.Controller;
 
 import javax.persistence.Column;
 import javax.persistence.OneToOne;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Size;
 import java.sql.Time;
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -25,18 +29,23 @@ public class CruiseDTO {
     private String nameOfCruise;
     @DateTimeFormat(pattern="yyyy-MM-dd")
     private LocalDate date;
+    
     private LocalTime time;
-    private int cost;
-    private int numberOfSeats;
-    private int duration;
+
+    private Long cost;
+    
+    private Long numberOfSeats;
+
+    private Long duration;
+
     private Integer route;
+
     private Integer ship;
     private String formError;
-
     public CruiseDTO() {
     }
 
-    public CruiseDTO(String nameOfCruise, LocalDate date, LocalTime time, int cost, int numberOfSeats, int duration, Integer route, Integer ship) {
+    public CruiseDTO(String nameOfCruise, LocalDate date, LocalTime time, Long cost, Long numberOfSeats, Long duration, Integer route, Integer ship) {
         this.nameOfCruise = nameOfCruise;
         this.date = date;
         this.time = time;
